requirements.txt
backcall==0.2.0
Click==7.0
decorator==5.0.9
Flask==1.0.3
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
ipython==7.24.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.10.1
MarkupSafe==1.1.1
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
Pygments==2.9.0
pytesseract==0.3.7
SQLAlchemy==1.3.3
traitlets==5.0.5
wcwidth==0.2.5
Werkzeug==2.0.1
